+++ example/lib/main.dart
@@ -1,11 +1,6 @@
-import 'dart:math' as math;
-import 'package:example/example/ex_text_border.dart';
-import 'package:example/example/ex_zo_dual_border.dart';
-import 'package:example/example/ex_zo_scribble_border.dart';
+import 'package:example/example/ex_zo_fire_border.dart';
+
 import 'package:flutter/material.dart';
-import 'package:characters/characters.dart';
-import 'package:zo_animated_border/painter/zo_rotating_text_border_painter.dart';
-import 'package:zo_animated_border/zo_animated_border.dart';
 
 void main() {
   runApp(const MyApp());
@@ -19,7 +14,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       theme: ThemeData.dark(),
-      home: const ExZoDualBorder(),
+      home: const ExZoFireBorder(),
     );
   }
 }